requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
hydra-core==1.3.2
idna==3.4
kaggle==1.5.16
keras==2.13.1
libclang==16.0.6
loguru==0.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opendatasets==0.1.22
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
